noreturn: snapshot wip of adding noreturn attribute
Currently Self code uses a very old gcc extension to mark noreturn
functions as "volatile void".  But these days it just triggers
-Wignored-qualifiers.  Thie commit switches to proper noreturn
attribute, but there's a problem with fatal_handler().

To begin with, there are two copies of it :) One (in lprintf.cpp) was
fixed to not take nested fatal errors in PR 115 back in 2017, the
other (in interruptedCtx.cpp) was not.

But when a function is declared noreturn, gcc takes this into account
at call site, so returning from (a nested) fatal_handler will
probably wreak havoc anyway.  Until this issue is settled I'd rather
stash this patch on a branch.