debug.hh: cast SPEND_TIME_FOR_DEBUGGING_BY_DEFAULT to bool
It is defined to 0 or 1 instead of false or true and I don't want to
search for and change all those other places (including cmake).  But a
wrong type here, integer, might cause warnings when this value is used
in a context where it must be boolean (though nothing in the code
currently does).  NFCI.